requirements.txt
paddleocr==2.6.1.3
easyocr==1.7.1
plotly==5.10.0
streamlit==1.30.0
pillow==9.5.0
numpy==1.24.4
pandas==2.0.3